chore: 목 데이터 상수 위치 변경
